% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_chartcode.R
\name{create_chartcode}
\alias{create_chartcode}
\title{Construct a single chart code from user input}
\usage{
create_chartcode(chartgrp = c("nl2010", "preterm", "who"),
  etn = c("nl", "tu", "ma", "hs"), sex = c("male", "female"),
  agegrp = c("0-15m", "0-4y", "1-21y", "0-21y", "0-4ya"),
  side = c("front", "back", "-hdc", "both", "hgt", "wgt", "hdc", "bmi",
  "wfh"), week = 32, language = c("dutch", "english"), version = "")
}
\arguments{
\item{chartgrp}{The chart group: \code{'nl2010'}, \code{'preterm'} or \code{'who'}}

\item{etn}{Either \code{'nl'}, \code{'tu'}, \code{'ma'} or
\code{'hs'}. May also be abbreviates to a single letter.}

\item{sex}{Either \code{'male'} or \code{'female'}}

\item{agegrp}{Either \code{'0-15m'}, \code{'0-4y'}, \code{'1-21y'}, \code{'0-21y'}
or \code{'0-4ya'}. Age group \code{'0-4ya'} provides the 0-4 chart with weight
for age (design E).}

\item{side}{Either \code{'front'}, \code{'back'}, \code{'-hdc'} or \code{'both'}}

\item{week}{A number between 25 and 36 gestational age. Only used for preterm charts.}

\item{language}{Language: \code{'dutch'} or \code{'english'}}

\item{version}{Version number. Default is to have no version number.}
}
\description{
Construct a single chart code from user input
}
