% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chartgrp.R
\name{get_chartgrp}
\alias{get_chartgrp}
\title{Obtain the chart group from the chartcode}
\usage{
get_chartgrp(chartcode = NULL)
}
\arguments{
\item{chartcode}{Chart codes, usually constructed by
\code{create_chartcode()}. If not specified, the function uses
all chart codes.}
}
\value{
The function returns a named vector with
  \code{length(chartcode)} elements with the chart group names.
}
\description{
Obtain the chart group from the chartcode
}
\details{
Warning: This function uses the property that the
  \code{chartgrp} can be uniquely be identified from the
  \code{chartcode}. When the same chartcode's are used in different
  chart groups, this function will give unpredictable results.
}
\examples{
get_chartgrp(c("NJAA", "NJBA"))
}
\seealso{
\code{\link{create_chartcode}}
}
