% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inverse_tx.R
\name{get_inverse_tx}
\alias{get_inverse_tx}
\title{Obtain the inverse tx() transformation function}
\usage{
get_inverse_tx(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
An function with one argument \code{y}
}
\description{
Obtain the inverse tx() transformation function
}
\examples{
get_inverse_tx("nl2010", "NJAA", "hgt")
}
