% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refcode.R
\name{get_refcode}
\alias{get_refcode}
\title{Obtain the name of the reference code used on a chart}
\usage{
get_refcode(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}}

\item{yname}{Name of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
\code{get_refcode()} returns a string.
}
\description{
Obtain the name of the reference code used on a chart
}
\examples{
get_refcode("NJAA", "hgt")
}
