% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq.R
\name{get_seq}
\alias{get_seq}
\title{Obtain the transformation sequence from a chartcode and yname}
\usage{
get_seq(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by
\code{create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
An string, either \code{"tr"} (transform-reference) or
  \code{"rt"} (reference-transform) specifying the transformation
  sequence.
}
\description{
Obtain the transformation sequence from a chartcode and yname
}
\examples{
get_ty("nl2010", "NJAA", "hgt")
}
