% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ty.R
\name{get_ty}
\alias{get_ty}
\title{Obtain the ty() transformation function}
\usage{
get_ty(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
An function with one argument \code{y}
}
\description{
Obtain the ty() transformation function
}
\examples{
get_ty("nl2010", "NJAA", "hgt")
}
