% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_viewport.R
\name{get_viewport}
\alias{get_viewport}
\title{Obtain the viewport name of the chart}
\usage{
get_viewport(chartcode, yname, chartgrp = get_chartgrp(chartcode),
  number = TRUE)
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}

\item{number}{If \code{number == TRUE} the function return the viewport
number. If \code{number == FALSE} the function return the viewport name.}
}
\value{
Number, or name, of the viewport
}
\description{
Obtain the viewport name of the chart
}
