% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ynames.R
\name{get_ynames}
\alias{get_ynames}
\title{Obtain response variables from a chartcode}
\usage{
get_ynames(chartcode = NULL, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}.
Can be a vector. If not specified, the function returns all
names.}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
A named vector with names of the response variables for
}
\description{
Obtain response variables from a chartcode
}
\examples{
get_ynames("NJAA")
}
