% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chartcode.R
\name{parse_chartcode}
\alias{parse_chartcode}
\title{Parses the chart code}
\usage{
parse_chartcode(chartcode = NULL)
}
\arguments{
\item{chartcode}{the chart code, usually constructed by \code{create_chartcode()}}
}
\value{
The function returns a \code{list} with the following components:
\describe{
   \item{\code{population}}{A string identifying the population,
   e.g. \code{'NL'},\code{'MA'}, \code{'TU'} or \code{'PT'}.}
   \item{\code{sex}}{A string \code{"male"}, \code{"female"} or
   \code{"undifferentiated"}.}
   \item{\code{design}}{A letter indicating the chart design: \code{'A'} = 0-15m,
   \code{'B'} = 0-4y, \code{'C'} = 1-21y, \code{'D'} = 0-21y, \code{'E'} =
   0-4ya.}
   \item{\code{side}}{A string indicating the side or \code{yname}:
   \code{'front'}, \code{'back'}, \code{'both'}, \code{'hgt'},
   \code{'wgt'}, \code{'hdc'}, \code{'bmi'}, \code{'wfh'}}
   \item{\code{language}}{The language in which the chart is drawn. Currently only
   \code{"dutch"} charts are implemented, but for \code{population == "PT"} we
   may also have \code{"english"}.}
   \item{\code{week}}{A scalar indicating the gestational age at birth.
   Only used if \code{population == "PT"}.}
   }
}
\description{
Parses the chart code
}
