% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breakpoints.R
\name{get_breakpoints}
\alias{get_breakpoints}
\title{Find table of break ages and labels}
\usage{
get_breakpoints(dnr = "smocc")
}
\arguments{
\item{dnr}{Name of donor dataset}
}
\value{
A data frame with three columns: \code{visit}, \code{age} and \code{label}
}
\description{
Find table of break ages and labels
}
\examples{
get_breakpoints("pops")
}
