% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference.R
\name{get_reference}
\alias{get_reference}
\title{Obtain the growth reference behind the chart}
\usage{
get_reference(chartcode, yname, chartgrp = get_chartgrp(chartcode))
}
\arguments{
\item{chartcode}{The chart code, returned by \code{create_chartcode()}}

\item{yname}{Names of the response variable}

\item{chartgrp}{The chart group. If not specified, it is calculated
automatically.}
}
\value{
An object of class \code{reference}
}
\description{
Obtain the growth reference behind the chart
}
\examples{
\dontrun{
get_reference("nl2010", "NJAA", "hgt")
}
}
\seealso{
\code{\link[clopus]{reference-class}}
}
